# contracts/openapi/replen-api.v1.yaml
openapi: 3.1.0
info:
  title: Replen planner API
  version: 1.0.0
  description: |
    Transactional API for the Replen ordering and replenishment platform (replen-api).
    All quantities are integer units. Money is GBP with two decimals.
    Errors use RFC 9457 problem details. Mutating POSTs accept an Idempotency-Key header.
servers:
  - url: http://localhost:4000
security:
  - bearer: []
...
paths:
...
  /api/v1/imports/{entity}:
    post:
      tags: [reference]
      summary: Upsert a batch of reference, inventory or history records (admin)
      parameters:
        - name: entity
          in: path
          required: true
          schema:
            type: string
            enum: [locations, suppliers, products, sourcing, item-locations, order-constraints, inventory-snapshots, open-purchase-orders, receipt-history]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [records]
              properties:
                records: { type: array, items: { type: object, additionalProperties: true } }
      responses:
        "200":
          description: Imported
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ImportResult" }
        "400": { $ref: "#/components/responses/Problem" }
        "403": { $ref: "#/components/responses/Problem" }
...
  /api/v1/inventory-positions:
    get:
      tags: [inventory]
      parameters:
        - { name: sku, in: query, schema: { type: string } }
        - { name: locationId, in: query, schema: { type: string } }
      responses:
        "200":
          description: Positions
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/InventoryPosition" } }
...
  /api/v1/purchase-orders/{poId}:
    get:
      tags: [purchasing]
      parameters:
        - { name: poId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Purchase order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PurchaseOrder" }
        "404": { $ref: "#/components/responses/Problem" }
  /api/v1/purchase-orders/{poId}/retry-submission:
    post:
      tags: [purchasing]
      parameters:
        - { name: poId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        "200":
          description: Submission retried
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PurchaseOrder" }
        "409": { $ref: "#/components/responses/Problem" }
  /api/v1/audit-events:
    get:
      tags: [audit]
      parameters:
        - { name: entityType, in: query, schema: { type: string } }
        - { name: entityId, in: query, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500 } }
      responses:
        "200":
          description: Audit events, newest first
          content:
            application/json:
              schema: { type: array, items: { $ref: "#/components/schemas/AuditEvent" } }
...
components:
  securitySchemes:
    bearer:
      type: http
      scheme: bearer
...
  responses:
    Problem:
      description: Problem details
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
    Problem:
      type: object
      required: [type, title, status]
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        code: { type: string }
        correlationId: { type: string }
        errors: { type: array, items: { type: object, additionalProperties: true } }
...
    ImportResult:
      type: object
      required: [entity, rows, batchId]
      properties:
        entity: { type: string }
        rows: { type: integer }
        batchId: { type: string, format: uuid }
...
    OpenOrder:
      type: object
      required: [reference, quantity, expectedDate, source]
      properties:
        reference: { type: string }
        quantity: { type: integer }
        expectedDate: { type: string, format: date }
        source: { type: string, enum: [LEGACY, REPLEN] }
    InventoryPosition:
      type: object
      required: [sku, locationId, onHand, reserved, inTransit, damaged, returnsPending, onOrder, availableToSell, inventoryPosition]
      properties:
        sku: { type: string }
        locationId: { type: string }
        asOfDate: { type: [string, "null"], format: date }
        onHand: { type: integer }
        reserved: { type: integer }
        inTransit: { type: integer }
        damaged: { type: integer }
        returnsPending: { type: integer }
        onOrder: { type: integer }
        availableToSell: { type: integer, description: "max(0, on-hand - reserved)" }
        inventoryPosition: { type: integer, description: "max(on-hand,0) - reserved + in-transit + on-order" }
        openOrders: { type: array, items: { $ref: "#/components/schemas/OpenOrder" } }
...
    PurchaseOrderLine:
      type: object
      required: [lineNo, sku, quantity, unitCost]
      properties:
        lineNo: { type: integer }
        sku: { type: string }
        quantity: { type: integer }
        unitCost: { type: number }
        receivedQty: { type: integer }
    PurchaseOrder:
      type: object
      required: [poId, poNumber, sourceProposalId, supplierId, destinationLocationId, status, totalValue, currency, lines, createdAt]
      properties:
        poId: { type: string, format: uuid }
        poNumber: { type: string }
        sourceProposalId: { type: string, format: uuid }
        supplierId: { type: string }
        destinationLocationId: { type: string }
        orderDate: { type: string, format: date }
        expectedDeliveryDate: { type: string, format: date }
        status: { type: string, enum: [CREATED, SUBMITTED, SUBMISSION_FAILED, CONFIRMED, PARTIALLY_RECEIVED, RECEIVED, CANCELLED] }
        erpPoNumber: { type: [string, "null"] }
        submissionAttempts: { type: integer }
        lastError: { type: [string, "null"] }
        totalValue: { type: number }
        currency: { type: string }
        approvedBy: { type: [string, "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        lines: { type: array, items: { $ref: "#/components/schemas/PurchaseOrderLine" } }
    AuditEvent:
      type: object
      required: [seq, eventId, occurredAt, actor, action, entityType, entityId, hash]
      properties:
        seq: { type: integer }
        eventId: { type: string, format: uuid }
        occurredAt: { type: string, format: date-time }
        actor: { type: string }
        action: { type: string }
        entityType: { type: string }
        entityId: { type: string }
        correlationId: { type: [string, "null"] }
        before: {}
        after: {}
        metadata: {}
        prevHash: { type: [string, "null"] }
        hash: { type: string }
...